# docs/swagger.yaml
swagger: "2.0"
info:
  title: Restfull API
  description: API hệ thống bán lẻ đồ gia dụng
  version: "1.0"
host: "26.107.218.28:5000"
basePath: "/api"
schemes:
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Token phải có dạng 'Bearer <JWT>'"
...
paths:
...
  /review/product/{product_id}:
    get:
      tags:
        - review
      summary: 'Lấy danh sách đánh giá của sản phẩm'
      parameters:
        - name: product_id
          in: path
          required: true
          type: integer
          description: 'ID của sản phẩm cần lấy đánh giá'
      responses:
        200:
          description: 'Danh sách đánh giá của sản phẩm'
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                user:
                  type: string
                rating:
                  type: integer
                comment:
                  type: string
                created_at:
                  type: string
                  format: date-time
        404:
          description: 'Không tìm thấy sản phẩm'
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Product not found'
  /review/user/{user_id}:
    get:
      tags:
        - review
      summary: 'Lấy danh sách đánh giá của người dùng'
      security:
        - Bearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
          description: 'ID của người dùng cần lấy đánh giá'
      responses:
        200:
          description: 'Danh sách đánh giá của người dùng'
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                product_id:
                  type: integer
                product_name:
                  type: string
                rating:
                  type: integer
                comment:
                  type: string
                created_at:
                  type: string
                  format: date-time
        403:
          description: 'Không có quyền truy cập'
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Forbidden: You can only view your own reviews'
        401:
          description: 'Chưa xác thực'
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Missing Authorization Header'
  /review/add-review/{product_id}:
    post:
      tags:
        - review
      summary: 'Thêm đánh giá sản phẩm (Chỉ Customer đã mua hàng)'
      security:
        - Bearer: []
      parameters:
        - name: product_id
          in: path
          required: true
          type: integer
          description: 'ID của sản phẩm cần đánh giá'
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              rating:
                type: integer
              comment:
                type: string
      responses:
        201:
          description: 'Đánh giá được thêm thành công'
        403:
          description: 'Chỉ có thể đánh giá sản phẩm đã mua'
        400:
          description: 'Lỗi dữ liệu đầu vào'
  /review/update/{review_id}:
    put:
      tags:
        - review
      summary: 'Sửa đánh giá (Chỉ người đã viết)'
      security:
        - Bearer: []
      parameters:
        - name: review_id
          in: path
          required: true
          type: integer
          description: 'ID của đánh giá cần sửa'
        - name:
          in: body
          required: true
          schema:
            type: object
            properties:
              rating:
                type: integer
              comment:
                type: string
      responses:
        200:
          description: 'Đánh giá được cập nhật thành công'
        403:
          description: 'Chỉ có thể sửa đánh giá của chính mình'
        404:
          description: 'Không tìm thấy đánh giá'
  /review/delete/{review_id}:
    delete:
      tags:
        - review
      summary: 'Xóa đánh giá (Chỉ người đã viết hoặc Admin)'
      security:
        - Bearer: []
      parameters:
        - name: review_id
          in: path
          required: true
          type: integer
          description: 'ID của đánh giá cần xóa'
      responses:
        200:
          description: 'Đánh giá đã được xóa'
        403:
          description: 'Chỉ có thể xóa đánh giá của chính mình hoặc là Admin'
        404:
          description: 'Không tìm thấy đánh giá'
  /review/reviews:
    get:
      tags:
        - review
      summary: 'Lấy toàn bộ đánh giá (Chỉ Admin)'
      security:
        - Bearer: []
      parameters:
        - name: product_id
          in: query
          required: false
          type: integer
          description: 'Lọc theo ID sản phẩm'
        - name: user_id
          in: query
          required: false
          type: integer
          description: 'Lọc theo ID người dùng'
        - name: page
          in: query
          required: false
          type: integer
          description: 'Trang số'
        - name: limit
          in: query
          required: false
          type: integer
          description: 'Số lượng đánh giá mỗi trang'
      responses:
        200:
          description: 'Danh sách đánh giá'
          schema:
            type: object
            properties:
              total:
                type: integer
              page:
                type: integer
              per_page:
                type: integer
              reviews:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    product_id:
                      type: integer
                    product_name:
                      type: string
                    user_id:
                      type: integer
                    user_name:
                      type: string
                    rating:
                      type: integer
                    comment:
                      type: string
                    created_at:
                      type: string
                      format: date-time
...